feat: Add Twitter Analytics tag
