+++ pages/_app.tsx
@@ -1,8 +1,7 @@
 import { AppProps } from 'next/app'
 import { useRouter } from 'next/router'
 import { useEffect } from 'react'
-import Script from 'next/script'
-
+import AnalyticsTags from '../components/analyticstags'
 import '../lib/global.css'
 
 export default function MyApp({ Component, pageProps }: AppProps) {
@@ -14,16 +13,7 @@ export default function MyApp({ Component, pageProps }: AppProps) {
   }, [router.query.ref])
   return (
   <>
-    <Script src="https://www.googletagmanager.com/gtag/js?id=AW-11328012362"/>
-    <Script id="google-analytics">
-    {`
-      window.dataLayer = window.dataLayer || [];
-      function gtag(){dataLayer.push(arguments);}
-      gtag('js', new Date());
-
-      gtag('config', 'AW-11328012362');
-    `}
-    </Script>
+    <AnalyticsTags />
     <Component {...pageProps} /> 
   </>
   )
